fibers

+++ runtime/vm/flag_list.h
@@ -242,9 +242,9 @@ constexpr bool FLAG_support_il_printer = false;
   R(eliminate_type_checks, true, bool, true,                                   \
     "Eliminate type checks when allowed by static type analysis.")             \
   D(support_rr, bool, false, "Support running within RR.")                     \
-  P(coroutines_registry_initial_size, size_t, 64,                              \
+  P(coroutines_registry_initial_size, int, 64,                                 \
     "All coroutines array initial size.")                                      \
-  P(coroutines_registry_shrink_marker, size_t, 1024 * 1024,                    \
+  P(coroutines_registry_shrink_marker, int, 1024 * 1024,                       \
     "All coroutines array initial size.")                                      \
   P(verify_entry_points, bool, false,                                          \
     "Throw API error on invalid member access through native API. See "        \
